// src/app/modules/staff/staff-admissions/model/staff-details-view.ts
/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { JobApplicationPersonView } from './job-application-person-view';
import { PassportView } from './passport-view';


/**
 * Gets or sets StaffDetailsView
 */
export interface StaffDetailsView { 
    /**
     * Gets Or Sets Id
     */
    id?: string;
    /**
     * Gets or sets StaffTypeId
     */
    staffTypeId?: number;
    /**
     * Gets or sets EmergencyContactName1
     */
    emergencyContactName1?: string;
    /**
     * Gets or sets EmergencyContactRelation1
     */
    emergencyContactRelation1?: string;
    /**
     * Gets or sets EmgergencyContactTelNumber1
     */
    emgergencyContactTelNumber1?: string;
    /**
     * Gets or sets EmergencyContactName2
     */
    emergencyContactName2?: string;
    /**
     * Gets or sets EmergencyContactRelation2
     */
    emergencyContactRelation2?: string;
    /**
     * Gets or sets EmgergencyContactTelNumber2
     */
    emgergencyContactTelNumber2?: string;
    /**
     * Gets or sets Position
     */
    position?: string;
    /**
     * Gets or sets Department
     */
    department?: string;
    /**
     * Gets or sets SpecialityQualification
     */
    specialityQualification?: string;
    /**
     * Gets or sets DateOfJoin
     */
    dateOfJoin?: Date;
    /**
     * Gets or sets MedicalCondition
     */
    medicalCondition?: string;
    /**
     * Gets or sets JobApplicationId
     */
    jobApplicationId?: string;
    /**
     * Gets or sets ProcessedComments
     */
    processedComments?: string;
    /**
     * Gets or sets EndDate
     */
    endDate?: Date;
    /**
     * Gets or sets Person
     */
    person?: JobApplicationPersonView;
    /**
     * Gets Or Sets Passport
     */
    passport?: PassportView;
}